+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/SearchHighlighter.java
@@ -55,7 +55,9 @@ public final class SearchHighlighter {
 	private final Value<String> searchStringValue = Value.nonNull("")
 					.listener(this::searchAndHighlightResults)
 					.build();
-	private final State caseSensitiveState = State.state();
+	private final State caseSensitiveState = State.builder()
+					.listener(this::searchAndHighlightResults)
+					.build();
 	private final Highlighter highlighter = new DefaultHighlighter();
 	private final List<MatchPosition> searchTextPositions = new ArrayList<>();
 	private final Value<Integer> currentSearchTextPositionIndex = Value.value();
@@ -216,7 +218,6 @@ public final class SearchHighlighter {
 	}
 
 	private void bindEvents(JTextComponent textComponent) {
-		caseSensitiveState.addListener(this::searchAndHighlightResults);
 		textComponent.getDocument().addDocumentListener((DocumentAdapter) e -> searchAndHighlightResults());
 		selectedSearchTextPosition.addConsumer(selectedSearchPosition -> {
 			if (selectedSearchPosition != null) {