+++ swing/framework-server-monitor/src/main/java/is/codion/swing/framework/server/monitor/ui/ClientInstanceMonitorPanel.java
@@ -147,8 +147,9 @@ public final class ClientInstanceMonitorPanel extends JPanel {
 						.build();
 		Font font = textArea.getFont();
 		textArea.setFont(new Font(Font.MONOSPACED, font.getStyle(), font.getSize()));
-		State lineWrapState = State.state();
-		lineWrapState.addConsumer(textArea::setLineWrap);
+		State lineWrapState = State.builder()
+						.consumer(textArea::setLineWrap)
+						.build();
 		textArea.setComponentPopupMenu(menu(Controls.builder()
 						.control(Control.builder(() -> saveLogToFile(textArea))
 										.name("Save to file..."))